+++ internal/webhook/stage/webhook.go
@@ -2,11 +2,9 @@ package stage
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
-	"github.com/Masterminds/semver"
-	"github.com/argoproj-labs/argocd-image-updater/pkg/image"
-	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -119,100 +117,20 @@ func (w *webhook) validateSubs(
 	if subs == nil { // nil subs is caught by declarative validations
 		return nil
 	}
-	// Can subscribe to repos XOR upstream Stages
-	if (subs.Repos == nil && len(subs.UpstreamStages) == 0) ||
-		(subs.Repos != nil && len(subs.UpstreamStages) > 0) {
+	// Can subscribe to Warehouse XOR upstream Stages
+	if (subs.Warehouse == "" && len(subs.UpstreamStages) == 0) ||
+		(subs.Warehouse != "" && len(subs.UpstreamStages) > 0) {
 		return field.ErrorList{
 			field.Invalid(
 				f,
 				subs,
 				fmt.Sprintf(
-					"exactly one of %s.repos or %s.upstreamStages must be defined",
+					"exactly one of %s.warehouse or %s.upstreamStages must be defined",
 					f.String(),
 					f.String(),
 				),
 			),
 		}
-	}
-	return w.validateRepoSubs(f.Child("repos"), subs.Repos)
-}
-
-func (w *webhook) validateRepoSubs(
-	f *field.Path,
-	subs *kargoapi.RepoSubscriptions,
-) field.ErrorList {
-	if subs == nil {
-		return nil
-	}
-	// Must subscribe to at least one repo of some sort
-	if len(subs.Git) == 0 && len(subs.Images) == 0 && len(subs.Charts) == 0 {
-		return field.ErrorList{
-			field.Invalid(
-				f,
-				subs,
-				fmt.Sprintf(
-					"at least one of %s.git, %s.images, or %s.charts must be non-empty",
-					f.String(),
-					f.String(),
-					f.String(),
-				),
-			),
-		}
-	}
-	errs := w.validateImageSubs(f.Child("images"), subs.Images)
-	return append(errs, w.validateChartSubs(f.Child("charts"), subs.Charts)...)
-}
-
-func (w *webhook) validateImageSubs(
-	f *field.Path,
-	subs []kargoapi.ImageSubscription,
-) field.ErrorList {
-	var errs field.ErrorList
-	for i, sub := range subs {
-		errs = append(errs, w.validateImageSub(f.Index(i), sub)...)
-	}
-	return errs
-}
-
-func (w *webhook) validateImageSub(
-	f *field.Path,
-	sub kargoapi.ImageSubscription,
-) field.ErrorList {
-	var errs field.ErrorList
-	if err := validateSemverConstraint(
-		f.Child("semverConstraint"),
-		sub.SemverConstraint,
-	); err != nil {
-		errs = field.ErrorList{err}
-	}
-	if sub.Platform != "" {
-		if _, _, _, err := image.ParsePlatform(sub.Platform); err != nil {
-			errs = append(errs, field.Invalid(f.Child("platform"), sub.Platform, ""))
-		}
-	}
-	return errs
-}
-
-func (w *webhook) validateChartSubs(
-	f *field.Path,
-	subs []kargoapi.ChartSubscription,
-) field.ErrorList {
-	var errs field.ErrorList
-	for i, sub := range subs {
-		errs = append(errs, w.validateChartSub(f.Index(i), sub)...)
-	}
-	return errs
-}
-
-func (w *webhook) validateChartSub(
-	f *field.Path,
-	sub kargoapi.ChartSubscription,
-) field.ErrorList {
-	if err := validateSemverConstraint(
-		f.Child("semverConstraint"),
-		sub.SemverConstraint,
-	); err != nil {
-		return field.ErrorList{err}
 	}
 	return nil
 }
@@ -309,19 +227,6 @@ func (w *webhook) validateHelmPromotionMechanism(
 				),
 			),
 		}
-	}
-	return nil
-}
-
-func validateSemverConstraint(
-	f *field.Path,
-	semverConstraint string,
-) *field.Error {
-	if semverConstraint == "" {
-		return nil
-	}
-	if _, err := semver.NewConstraint(semverConstraint); err != nil {
-		return field.Invalid(f, semverConstraint, "")
 	}
 	return nil
 }
